Fix warnings in vs2022

+++ src/Environment/Global/ClockGenerator.h
@@ -23,5 +23,5 @@ class ClockGenerator {
 
  private:
   std::vector<ITickable*> components_;
-  unsigned long long timer_count_;
+  int timer_count_; // TODO consider size, unsigned
 };
